QLDA: inline lọc phòng ban đơn vị, bỏ DmDonViPhongBanScope
Gộp filter trực tiếp vào query/import handler; DanhMucDonViGetDanhSachQuery dùng WhereFunc thay if-block.
