+++ QLDA.Application/KeHoachTrienKhaiHangMuc/KeHoachTrienKhaiHangMucImportUserScope.cs
@@ -1,9 +1,17 @@
 using BuildingBlocks.Domain.Providers;
-using QLDA.Application.DanhMucDonVis;
 
 namespace QLDA.Application.KeHoachTrienKhaiHangMucs;
 
 internal static class KeHoachTrienKhaiHangMucImportUserScope {
-    internal static long? TryGetCurrentDonViId(IUserProvider userProvider) =>
-        DmDonViPhongBanScope.TryGetCurrentDonViId(userProvider);
+    internal static long? TryGetCurrentDonViId(IUserProvider userProvider) {
+        if (userProvider.Id <= 0)
+            return null;
+
+        try {
+            var donViId = userProvider.Info.DonViID;
+            return donViId > 0 ? donViId : null;
+        } catch (UnauthorizedAccessException) {
+            return null;
+        }
+    }
 }
